requirements.txt
streamlit
tf-keras
tensorflow
pillow
numpy
streamlit-webrtc
av